Adjust number of average blocks since the round length is 32.768s now
I think 220 should be fine since the progress bar has a precision of 34 bars only. Currently the progress bar is at 100% already after 110 blocks.

Time based version to estimate the current era based on the number of blocks of the previous era:
https://github.com/sacherjj/casper-network-monitor/blob/clean-up-refactor/app/cnm/data/abstract_era_times.py

Explanation from the author:
--- START ---
I schedule a task to run 2 hours and 5 minutes after the last to get the latest time and walk them forward as estimates.  So a simple time of first block of current Era plus 2 hours and 9 seconds is the expected time of next era.
--- END ---